pyproject.toml
[tool.poetry]
name = "alist"
version = "0.1.0"
description = ""
authors = ["adoutt <649384645@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
